Se ha agregado las validaciones a todos los campos y se ha terminado los requerimientos funcionales en el modulo de inventario

+++ src/app/componentes/inventario/inventario.component.ts
@@ -6,9 +6,9 @@ import { NgForm } from '@angular/forms';
 import { Articulo } from 'src/app/Modelos/articulo';
 import * as $ from 'jquery';
 import { VencimientoArticulo } from 'src/app/Modelos/vencimiento-articulo';
+import Swal from 'sweetalert2';
+import Popper from 'popper.js';
 
-
- 
 @Component({
   selector: 'app-inventario',
   templateUrl: './inventario.component.html',
@@ -20,6 +20,7 @@ export class InventarioComponent implements OnInit {
     private auth: AngularFireAuth) {
 
   }
+
 
   //Atributos para manejo de articulos
   dataArticulo: Articulo[];
@@ -38,7 +39,7 @@ export class InventarioComponent implements OnInit {
     iva: true,
     descuento: 0,
     precio: 0,
-    precioVenta:0
+    precioVenta: 0
   }
   vencimientoSelected = {
     id: "",
@@ -53,15 +54,71 @@ export class InventarioComponent implements OnInit {
   searchData: Articulo[];
   search: string = "";
 
+  /*Auxiliares de Botones */
   modificarGuardar: string = "Editar";
   cancelarCerrar = "Cerrar";
-  statusEdit = false;
+  statusEdit:boolean = false;
+  mostrarFormulario:boolean=false;
 
-  prueba(algo?: any) {
+  /* Validadores de Formulario */
+
+  codigo: boolean = false;
+  precioCompra: boolean = false;
+  precioVenta: boolean = false;
+  descuento: boolean = false;
+  min: boolean = false;
+  max: boolean = false;
+  cantidad: boolean = false;
+  formValido: boolean = false;
+
+  codigoEdit: boolean = false;
+  precioCompraEdit: boolean = false;
+  precioVentaEdit: boolean = false;
+  descuentoEdit: boolean = false;
+  minEdit: boolean = false;
+  maxEdit: boolean = false;
+  formValidoEdit: boolean = false;
+
+  cantidadAdd:boolean=false;
+  formAddValid:boolean=false;
+
+
+
+  prueba() {
+
+    /*    Swal.fire({
+      title: 'Error!',
+      text: 'Do you want to continue',
+      icon: 'error',
+      confirmButtonText: 'Cool'
+    })  */
+
   }
+
+  mensajeError(elemento: string, mensaje: string) {
+    $("#" + mensaje).show();
+    var pop = new Popper(document.getElementById(elemento), document.getElementById(mensaje),
+      {
+        placement: "bottom",
+        modifiers: {
+          applyStyle: { enabled: true },
+          applyReactStyle: {
+            enabled: true,
+            order: 900,
+          },
+        }
+      }
+    );
+
+    setTimeout(() => {
+      $("#" + mensaje).hide();
+    }, 4000);
+  }
+
   ngOnInit() {
     this.loadArticle();
     this.loadVencimientoArticle();
+    $('#formRegister').hide();
 
   }
   singOut() {
@@ -101,62 +158,90 @@ export class InventarioComponent implements OnInit {
   }
   //Manipulacion de Articulos
   registerArticle(form: NgForm) {
-    this.inventarioServ.registerArticle(this.auth.auth.currentUser.email, form).then(
-      exito => {
-        this.toas.success('Se ha Registrado el Articulo', 'Operacion Exitosa');
-        //Registra la fecha de vencimiento
-        this.inventarioServ.registerVencimientoArticle(this.auth.auth.currentUser.email, form).then(
+    if (form.valid) {
+      if (this.formValido) {
+        this.inventarioServ.registerArticle(this.auth.auth.currentUser.email, form).then(
           exito => {
-            this.toas.success('Fechas Agregadas', 'Vencimiento');
-            this.clearForm(form);
+            Swal.fire('Registro Exitoso', 'Se ha agregado el articulo a la lista', 'success');
+            //Registra la fecha de vencimiento
+            this.inventarioServ.registerVencimientoArticle(this.auth.auth.currentUser.email, form).then(
+              exito => {
+                this.clearForm(form);
+              },
+              error => {
+                this.toas.warning('No se ha podido Registrar la fecha de Vencimiento', 'Fallo');
+                console.log(error);
+              }
+            )
 
           },
           error => {
-            this.toas.warning('No se ha podido Registrar la fecha de Vencimiento', 'Fallo');
+            this.toas.warning('No se ha podido Registrar', 'Fallo la Operacion');
             console.log(error);
           }
         )
-
-      },
-      error => {
-        this.toas.warning('No se ha podido Registrar', 'Fallo la Operacion');
-        console.log(error);
+      } else {
+        Swal.fire('Error en el Formulario', 'Verifique la informacion suministrada', 'error');
       }
-    )
+
+    }
+    else {
+      Swal.fire('Error en el Formulario', 'Verifique la informacion suministrada', 'error');
+
+    }
+
   }
   agregarInventario(form: NgForm) {
-    this.inventarioServ.registerVencimientoArticle(this.auth.auth.currentUser.email, form).then(
-      exito => {
-        let codigo = form.value.codigo;
-        this.toas.success('Se han agregado los nuevos articulos al inventario', 'Registro Exitoso');
-        this.fechasVencimientoAsociadas(codigo);
-        this.articuloSelected.codigo = codigo;
-      },
-      error => {
-        this.toas.warning('no se han podido agregar los nuevos articulos al inventario', 'Operacion Fallida');
-        console.log(error);
+    if(form.valid){
+      if(this.formAddValid){
+        this.inventarioServ.registerVencimientoArticle(this.auth.auth.currentUser.email, form).then(
+          exito => {
+            let codigo = form.value.codigo;
+            Swal.fire('Exito','Se ha registrado los nuevos productos al inventario','success');
+            this.fechasVencimientoAsociadas(codigo);
+            this.articuloSelected.codigo = codigo;
+          },
+          error => {
+            Swal.fire('Error','No se ha podido agregar los articulos al inventario','error');
+            console.log(error);
+          }
+        )
+      }else{
+        Swal.fire('Error','Verifica los campos resaltados','error');
       }
-    )
+    }else{
+      Swal.fire('Error','Debes llenar los campos correctamente','error');
+    }
   }
   viewArticle(article: any) {
     this.fechasVencimientoAsociadas(article.codigo);
     console.log(article);
     this.articuloSelected = article;
-   //Estado por defecto del Modal View, evita errores al hacer click fuera del modal
-      this.modificarGuardar = "Modificar";
-      this.cancelarCerrar = "Cerrar";
-      this.statusEdit = false;
-      $('#descripcion').attr('readonly', 'readonly');
-      $('#presentacion').attr('readonly', 'readonly');
-      $('#precio').attr('readonly', 'readonly');
-      $('#precioVenta').attr('readonly', 'readonly');
+    //Estado por defecto del Modal View, evita errores al hacer click fuera del modal
+    this.modificarGuardar = "Modificar";
+    this.cancelarCerrar = "Cerrar";
+    this.statusEdit = false;
+    $('#descripcion').attr('readonly', 'readonly');
+    $('#presentacion').attr('readonly', 'readonly');
+    $('#precio').attr('readonly', 'readonly');
+    $('#precioVenta').attr('readonly', 'readonly');
+    $('#descuento').attr('readonly', 'readonly');
+    $('#stockMinimo').attr('readonly', 'readonly');
+    $('#stockMaximo').attr('readonly', 'readonly');
+    $('#saveModif').attr("data-dismiss", "modal");
+    $('#saveModif').attr("data-dismiss", "modal");
+    $('#cancelar').attr("data-dismiss", "modal");
+    $('#btnFechas').show();
+    $('#btnDelete').show();
+    this.codigoEdit = false;
+    this.precioCompraEdit= false;
+    this.precioVentaEdit = false;
+    this.descuentoEdit = false;
+    this.minEdit = false;
+    this.maxEdit = false;
+    this.formValidoEdit = false;
 
-      $('#descuento').attr('readonly', 'readonly');
-      $('#stockMinimo').attr('readonly', 'readonly');
-      $('#stockMaximo').attr('readonly', 'readonly');
-      $('#saveModif').attr("data-dismiss", "modal");
-      
-      
+
 
   }
 
@@ -165,44 +250,48 @@ export class InventarioComponent implements OnInit {
   }
 
   deleteArticle(id: string) {
-    this.inventarioServ.deleteArticle(this.auth.auth.currentUser.email, id).then(
-      resultado => {
-        this.toas.info('Se ha Eliminado el Articulo', 'Eliminacion Exitosa');
-      }
-      ,
-      error => {
-        this.toas.error('No se ha podido Eliminar el articulo', 'Fallo al Eliminar');
-        console.log(error);
-      }
-    );
-    for(let i = 0; i< this.listVencimiento.length; i++){
-      this.inventarioServ.deleteVencimientoArticle(this.auth.auth.currentUser.email,this.listVencimiento[i].id).then(
-        success => {
-          this.toas.success('Fechas Asociadas','Eliminacion');
-        },
-        fail => {
-          this.toas.error('Fechas Asociadas','Error');
-          console.log(fail);
+    if(confirm('El Articulo sera Eliminado al igual que el Registro de Inventario Asociado')){
+      {
+        this.inventarioServ.deleteArticle(this.auth.auth.currentUser.email, id).then(
+          resultado => {
+            Swal.fire('Exito','El Articulo se ha eliminado correctamente','success');
+          }
+          ,
+          error => {
+            this.toas.error('No se ha podido Eliminar el articulo', 'Fallo al Eliminar');
+            console.log(error);
+          }
+        );
+        for (let i = 0; i < this.listVencimiento.length; i++) {
+          this.inventarioServ.deleteVencimientoArticle(this.auth.auth.currentUser.email, this.listVencimiento[i].id).then(
+            success => {
+              this.toas.info('Fechas Asociadas', 'Eliminacion');
+            },
+            fail => {
+              this.toas.error('fallo al eliminar las Fechas Asociadas', 'Error');
+              console.log(fail);
+            }
+          )
         }
-      )   
+    
+      }
     }
-  
-    }
-  deleteMultiArticle(){
+  }
+  deleteMultiArticle() {
 
   }
-  deleteVencimientoArticle(id:string,cod:string){
-    this.inventarioServ.deleteVencimientoArticle(this.auth.auth.currentUser.email,id).then(
+  deleteVencimientoArticle(id: string, cod: string) {
+    this.inventarioServ.deleteVencimientoArticle(this.auth.auth.currentUser.email, id).then(
       success => {
-        this.toas.success('Fecha Asociada','Eliminacion');
+        this.toas.success('Fecha Asociada', 'Eliminacion');
         this.loadVencimientoArticle();
         this.fechasVencimientoAsociadas(cod);
       },
       fail => {
-        this.toas.error('Fecha Asociada','Error');
+        this.toas.error('Fecha Asociada', 'Error');
         console.log(fail);
       }
-    )  
+    )
   }
 
   clearSearch() {
@@ -249,42 +338,59 @@ export class InventarioComponent implements OnInit {
       $('#saveModif').removeAttr("data-dismiss");
       this.toas.info('Modifique los campos deseados', 'Modificar');
       $('#saveModif').removeAttr("data-dismiss");
+/*       $('#cancelar').removeAttr("data-dismiss");
+ */      $('#btnFechas').hide();
+      $('#btnDelete').hide();
+
+
 
 
     }
     else {
-      this.modificarGuardar = "Modificar";
-      this.cancelarCerrar = "Cerrar";
-      this.statusEdit = false;
-      $('#descripcion').attr('readonly', 'readonly');
-      $('#presentacion').attr('readonly', 'readonly');
-      $('#precio').attr('readonly', 'readonly');
-      $('#precioVenta').attr('readonly','readonly');
-      $('#descuento').attr('readonly', 'readonly');
-      $('#stockMinimo').attr('readonly', 'readonly');
-      $('#stockMaximo').attr('readonly', 'readonly');
-      $('#saveModif').attr("data-dismiss", "modal");
-
-
-      //llamando al metodo de actualizacion de los cambios
-      this.inventarioServ.editArticle(form,this.auth.auth.currentUser.email).then(
-        exito => {
-            this.toas.show('se estan guardando los cambios', 'Modificando');
-            console.log(exito);
-            this.toas.success('Cambios guardados correctamente','Exito');
-        },
-        fail => {
-          this.toas.warning ('No se ha podido modificar','Operacion Fallida');
-          console.log(fail);
-
+      if(form.valid){
+        if(this.formValidoEdit){
+              this.modificarGuardar = "Modificar";
+              this.cancelarCerrar = "Cerrar";
+              this.statusEdit = false;
+              $('#descripcion').attr('readonly', 'readonly');
+              $('#presentacion').attr('readonly', 'readonly');
+              $('#precio').attr('readonly', 'readonly');
+              $('#precioVenta').attr('readonly', 'readonly');
+              $('#descuento').attr('readonly', 'readonly');
+              $('#stockMinimo').attr('readonly', 'readonly');
+              $('#stockMaximo').attr('readonly', 'readonly');
+              $('#saveModif').attr("data-dismiss", "modal");
+              $('#cancelar').attr("data-dismiss", "modal");
+              $('#btnFechas').show();
+              $('#btnDelete').show();
+        
+              //llamando al metodo de actualizacion de los cambios
+              this.inventarioServ.editArticle(form, this.auth.auth.currentUser.email).then(
+                exito => {
+                  Swal.fire('Exito','Se ha modificado el Articulo','success')
+                },
+                fail => {
+                  Swal.fire('Error','No se ha podido modificar el Articulo','error');
+                  console.log(fail);
+        
+                }
+              )
+        
+            
+          
+        }else{
+          Swal.fire('Error','has alterado algun campo de manera incorrecte, el campo aparecera con un mensaje debajo','error');
         }
-      )
+      }else{
 
+        Swal.fire('Error','Verifica los datos del formulario, puede que tengas un campo invalido','error');
+
+      }
     }
   }
   cancelCerrar() {
     if (this.statusEdit) {
-      this.modificarGuardar = "Editar";
+      this.modificarGuardar = "Modificar";
       this.cancelarCerrar = "Cerrar";
       this.statusEdit = false;
       $('#descripcion').attr('readonly', 'readonly');
@@ -294,18 +400,30 @@ export class InventarioComponent implements OnInit {
       $('#stockMinimo').attr('readonly', 'readonly');
       $('#stockMaximo').attr('readonly', 'readonly');
       $('#saveModif').removeAttr("data-dismiss");
+/*       $('#cancelar').removeAttr("data-dismiss");
+ */      $('#btnFechas').show();
+      $('#btnDelete').show();
       this.loadArticle();
       this.loadVencimientoArticle();
+      this.codigoEdit = false;
+      this.precioCompraEdit= false;
+      this.precioVentaEdit = false;
+      this.descuentoEdit = false;
+      this.minEdit = false;
+      this.maxEdit = false;
+      this.formValidoEdit = false;
+
 
     }
     else {
       this.statusEdit = false;
+      $('#cancelar').attr("data-dismiss", "modal");
       this.loadArticle();
       this.loadVencimientoArticle();
 
     }
   }
-  
+
 
   //Filtro de Vencimientos
   fechasVencimientoAsociadas(cod: string) {
@@ -329,14 +447,197 @@ export class InventarioComponent implements OnInit {
 
   }
 
+  /* metodo de validacion del formulario */
+
+
   //Metodos de Apoyo
+
   clearForm(form: NgForm) {
     form.reset();
+    this.codigo = false;
+    this.precioCompra = false;
+    this.precioVenta = false;
+    this.descuento = false;
+    this.min = false;
+    this.max = false;
+    this.cantidad = false;
+    this.formValido = false;
   }
 
   excentoIVA(e) {
     this.articuloSelected.iva = e.target.checked;
   }
+
+  validarForm(form: NgForm) {
+    if (form.value.codigo != undefined) {
+      if (this.validarNumero(form.value.codigo)) {
+        this.codigo = false;
+      } else {
+        this.codigo = true;
+      }
+    }
+    if (form.value.precio != undefined) {
+      if (this.validarNumero(form.value.precio)) {
+        this.precioCompra = false;
+      } else {
+        this.precioCompra = true;
+
+      }
+    }
+    if (form.value.precioVenta != undefined) {
+      if (this.validarNumero(form.value.precioVenta)) {
+        this.precioVenta = false;
+      } else {
+        this.precioVenta = true;
+      }
+    }
+    if (form.value.descuento != undefined) {
+      if (this.validarNumero(form.value.descuento)) {
+        this.descuento = false;
+      }
+      else {
+        this.descuento = true;
+
+      }
+    }
+    if (form.value.stockMinimo != undefined) {
+      if (this.validarNumero(form.value.stockMinimo)) {
+        this.min = false;
+      }
+      else {
+        this.min = true;
+
+      }
+    }
+    if (form.value.stockMaximo != undefined) {
+      if (this.validarNumero(form.value.stockMaximo)) {
+        this.max = false;
+      } else {
+        this.max = true;
+      }
+    }
+    if (form.value.cantidad != undefined) {
+      if (this.validarNumero(form.value.cantidad)) {
+        this.cantidad = false;
+      }
+      else {
+        this.cantidad = true;
+      }
+    }
+
+    if (!this.codigo && !this.precioCompra && !this.precioVenta && !this.min && !this.max && !this.descuento && !this.cantidad) {
+      this.formValido = true;
+    } else {
+      this.formValido = false;
+    }
+  }
+
+  validarFormEdit(form: NgForm) {
+    if (form.value.codigo != undefined) {
+      if (this.validarNumero(form.value.codigo)) {
+        this.codigoEdit = false;
+      } else {
+        this.codigoEdit = true;
+      }
+    }
+    if (form.value.precio != undefined) {
+      if (this.validarNumero(form.value.precio)) {
+        this.precioCompraEdit = false;
+      } else {
+        this.precioCompraEdit = true;
+
+      }
+    }
+    if (form.value.precioVenta != undefined) {
+      if (this.validarNumero(form.value.precioVenta)) {
+        this.precioVentaEdit = false;
+      } else {
+        this.precioVentaEdit = true;
+      }
+    }
+    if (form.value.descuento != undefined) {
+      if (this.validarNumero(form.value.descuento)) {
+        this.descuentoEdit = false;
+      }
+      else {
+        this.descuentoEdit = true;
+
+      }
+    }
+    if (form.value.stockMinimo != undefined) {
+      if (this.validarNumero(form.value.stockMinimo)) {
+        this.minEdit = false;
+      }
+      else {
+        this.minEdit = true;
+
+      }
+    }
+    if (form.value.stockMaximo != undefined) {
+      if (this.validarNumero(form.value.stockMaximo)) {
+        this.maxEdit = false;
+      } else {
+        this.maxEdit = true;
+      }
+    }
+    if (!this.codigoEdit && !this.precioCompraEdit && !this.precioVentaEdit && !this.minEdit && !this.maxEdit && !this.descuentoEdit) {
+      this.formValidoEdit = true;
+    } else {
+      this.formValidoEdit = false;
+    }
+  }
+  validarFormAdd(form:NgForm){
+    if(form.value.cantidad!=undefined){
+      if(this.validarNumero(form.value.cantidad)){
+        this.cantidadAdd=false;
+      }else{
+        this.cantidadAdd=true;
+      }
+  }
+  if(!this.cantidadAdd){
+    this.formAddValid=true;
+  }else{
+    this.formAddValid=false;
+  }
+}
+
+
+  validarNumero(valor: string) {
+    if (!/^([0-9])*$/.test(valor)) {
+      return false;
+    } else {
+      return true;
+    }
+
+  }
+  validarString(valor: string) {
+    if (!/^([a-z])*$/.test(valor.toLowerCase())) {
+      return false;
+    }
+    else {
+      return true;
+    }
+
+
+  }
+  swichAgregar() {
+    if (!this.mostrarFormulario) {
+      $('#formRegister').show();
+      $('#contenedorIcon').hide();
+      this.mostrarFormulario=true;
+      $('#tableArticulos').removeClass('col-md-10');
+
+    }
+    else {
+      $('#formRegister').hide();
+      $('#contenedorIcon').show();
+      this.mostrarFormulario=false;
+      $('#tableArticulos').addClass('col-md-10');
+
+    }
+  }
+
+
 
 
 }
